pyproject.toml
[project]
name = "llm-http-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "litestar[standard]>=2.16.0",
    "python-dotenv>=1.1.0",
]
